+++ lib/blurred_overlay.dart
@@ -1,7 +1,9 @@
 /// A Flutter package to show blurred dialogs and bottom sheets
 /// using [BackdropFilter] with optional shadow, rounded corners,
 /// and safe area handling.
+// ignore: unnecessary_library_name
 library blurred_overlay;
 
 export 'src/blurred_modal_bottom_sheet.dart';
 export 'src/blurred_dialog.dart';
+export 'src/blurred_drawer.dart';